Refactor to make Game the 'Source of Truth' for the score
Part of this refactoring involved having the Game.play(move:) method return a tuple consisting of a Boolean to indicate whether the move was correct or not and the game score after the move.

In addition, a couple more tests were added to GameTests.swift to validate that the response from the Game.play(move:) method returned non-nil values in the returned tuple.